Fix -Wconversion build breaks; repaint ComboBox indicator on accent change
Tighten up a few narrowing/sign conversions caught by -Wconversion -Werror, and fix VcComboBox's Canvas indicator so it repaints when the active accent colour changes (with a QML smoke test to prevent regressions).

+++ modules/ui/colour/oklchcolour.cpp
@@ -197,7 +197,11 @@ QColor toSrgbFitted( const Oklch &oklch )
     const double g = srgbEncode( linear.g );
     const double bl = srgbEncode( linear.b );
 
-    return QColor::fromRgbF( clamp01( r ), clamp01( g ), clamp01( bl ), 1.0 );
+    return QColor::fromRgbF(
+        static_cast<float>( clamp01( r ) ),
+        static_cast<float>( clamp01( g ) ),
+        static_cast<float>( clamp01( bl ) ),
+        1.0f );
 }
 
 Oklch fromSrgb( const QColor &srgb )
